Finish Main.java and Tester.java according to prof
Finish Main.java as per the spec of the professor. Main now uses only
one loop over the riskFactor array to display the stats of Stocks, ETFs,
and lastly both combined.

Finish Tester.java as per the spec of the professor. Tester now has the
correct run method, which was of worry beforehand.