Edit Helper.java file for purposes of Statistics
Edited the computeStats method so that it may calculate the proper
information needed from the averageProfitLong, averageProfitShort, and
averageProfit in Statistics. This was done by simply multiplying the
profit by 100 before adding it to totalProfitLong or totalProfitShort
in each trade.

Added the roundTwoDec method to round each attribute of Statistics to
two significant figures so that they may be better understood by the
viewer once displayed from the Main.java file. It does this in the
computeStats method before assigning each value to the appropriate
attribute in its Statistics object.